【cleancode】修复cleancode

Created-by: MemScopeDeveloper

Match-id-19b1a416296bf845ff32a241e4e1aab362d36c14

+++ csrc/analysis/memory_state_manager.h
@@ -51,7 +51,7 @@ public:
     explicit MemoryState(std::shared_ptr<MemoryEvent>& event)
     {
         events.push_back(event);
-        size = event->size;
+        size = static_cast<uint64_t>(event->size);
         leaksDefinedOwner = "";
         userDefinedOwner = event->describeOwner;
         inefficientType = "";
